backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
google-generativeai==0.3.2
langchain==0.1.10
langchain-community==0.0.27
langchain-google-genai==0.0.9
pinecone-client==3.1.0
scikit-learn==1.4.0
pandas==2.2.0
shap==0.44.1
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
pypdf==4.0.1
langchain-pinecone
pdfplumber
python-jose[cryptography]
requests
python-multipart
shap==0.44.1